feat(controller): update controller
